--- LocalPackage/Sources/Features/HouseworkFeature/HouseworkBoardView/SubViews/HouseworkDateHeaderContent.swift
@@ -1,48 +1,0 @@
-//
-//  HouseworkDateHeaderContent.swift
-//  homete
-//
-//  Created by 佐藤汰一 on 2025/09/06.
-//
-
-import HometeResources
-import HometeUI
-import SwiftUI
-
-struct HouseworkDateHeaderContent: View {
-    
-    @Environment(\.calendar) var calendar
-
-    @Binding var selectedDate: Date
-    
-    var body: some View {
-        HStack {
-            Button {
-                updateSelectedDate(value: -1)
-            } label: {
-                Image(systemName: "chevron.left")
-                    .foregroundStyle(.onSurface)
-            }
-            Spacer()
-            Text(selectedDate, format: .dateTime.day().month(.abbreviated).year())
-            Spacer()
-            Button {
-                updateSelectedDate(value: 1)
-            } label: {
-                Image(systemName: "chevron.right")
-                    .foregroundStyle(.onSurface)
-            }
-        }
-    }
-}
-
-private extension HouseworkDateHeaderContent {
-
-    func updateSelectedDate(value: Int) {
-
-        withAnimation {
-
-            selectedDate = calendar.date(byAdding: .day, value: value, to: selectedDate) ?? selectedDate
-        }
-    }
-}
