+++ LocalPackage/Sources/HometeDomain/Utilities/Preview/PreviewEnvironment.swift
@@ -33,26 +33,3 @@ public extension View {
             .environment(\.calendar, .japanese)
     }
 }
-
-public extension Date {
-    static func previewDate(
-        year: Int,
-        month: Int,
-        day: Int,
-        hour: Int = .zero,
-        minute: Int = .zero,
-        second: Int = .zero
-    ) -> Date {
-        DateComponents(
-            calendar: .japanese,
-            timeZone: .tokyo,
-            year: year,
-            month: month,
-            day: day,
-            hour: hour,
-            minute: minute,
-            second: second
-        )
-        .date! // swiftlint:disable:this force_unwrapping
-    }
-}
